ES1-1250,ES1-1224 : Live no progress events , CDVR recordings error
Reason for change:

Realtek will provide the position in absolute format when the qtdemux override (PTS restamping) is disabled.
However, if the first PTS value isn't sent to the pipeline, the segment event query will return 0.
Consequently, the player won't be able to calculate the position during tuning when the first PTS is non-zero.
The following code block aids in obtaining the correct segment start value for the scenario where QTDEMUX is disabled, specifically for ES1 and in the mac simulator.

Test Procedure: Test case updated in the ticket
Risks: low

DEPS_BRANCH=rdk/components/generic/gst-plugins-rdk-aamp:dev_sprint_24_2
DEPS_BRANCH=rdk/components/generic/aampabr:dev_sprint_24_2
DEPS_BRANCH=rdk/components/generic/aampmetrics:dev_sprint_24_2

+++ aampgstplayer.h
@@ -200,9 +200,9 @@ public:
          * @fn GetPositionMilliseconds
          * @retval playback position in MS
          */
-	long GetPositionMilliseconds(void) override;
-        /**
-         * @fn GetDurationMilliseconds 
+	long long GetPositionMilliseconds(void) override;
+	/**
+         * @fn GetDurationMilliseconds
          * @retval playback duration in MS
          */
 	long GetDurationMilliseconds(void) override;
